requirements.txt
Django==4.1.5
Pillow==9.4.0
PyYAML==6.0
black==22.12.0
boto3==1.26.44
celery==5.2.7
dj-database-url==1.2.0
dj-rest-auth==2.2.6
django-allauth==0.52.0
django-anymail==9.0
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
factory-boy==3.2.1
gunicorn==20.1.0
python-dotenv==0.21.0
uritemplate==4.1.1
